refactor(dashboard): lint fix
