+++ apps/dashboard/src/components/footer/FooterBar.vue
